+++ tests/end_to_end/test_app.py
@@ -201,16 +201,3 @@ def test_app_critical_only(
     # With critical only, the 4 single-source models (is_critical=false) should be skipped
     # so fewer forecasts than test_app
     assert db_session.query(ForecastSQL).count() > init_n_forecasts
-
-
-def test_get_all_models_critical_only():
-    """Test that get_critical_only filters to only critical models."""
-    from site_forecast_app.models.pydantic_models import get_all_models
-
-    all_models = get_all_models(client_abbreviation="nl")
-    critical_models = get_all_models(client_abbreviation="nl", get_critical_only=True)
-
-    assert len(critical_models.models) > 0
-    assert len(critical_models.models) < len(all_models.models)
-    assert all(m.is_critical for m in critical_models.models)
-    assert not all(m.is_critical for m in all_models.models)
