+++ tests/unit/models/test_pydantic_models.py
@@ -15,3 +15,15 @@ def test_site_group_uuid():
     model.site_group_uuid = "some-uuid"
 
     model = Model(**model.model_dump())
+
+def test_get_all_models_critical_only():
+    """Test that get_critical_only filters to only critical models."""
+    from site_forecast_app.models.pydantic_models import get_all_models
+
+    all_models = get_all_models(client_abbreviation="nl")
+    critical_models = get_all_models(client_abbreviation="nl", get_critical_only=True)
+
+    assert len(critical_models.models) > 0
+    assert len(critical_models.models) < len(all_models.models)
+    assert all(m.is_critical for m in critical_models.models)
+    assert not all(m.is_critical for m in all_models.models)
